google_patent/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
greenlet==3.1.0
numpy==2.1.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
six==1.16.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1